requirements.txt
discord.py[voice]==2.4.0
aiohttp>=3.9,<4
PyNaCl>=1.5,<2
